Terminate iteration through the hits based on count
Count the number of hits we've processed, and stop iteration when we
reach the end.  Without this, we sometimes get an error from
elasticsearch when dumping an
index because elasticsearch loses the scroll ID at the end of the
iteration.

This may only be a problem with old elasticsearches: I've
observed it with elasticsearch 0.90.12.  However, the fix should work
fine with all versions of elasticsearch.  The error looks something like:

    SearchContextMissingException[No search context found for id [....]];