Add restore_alias, for atomic swapping with aliases
So that we can switch a running aliased index for one in production,
this adds a `restore_alias` command, which functions similarly to the
`restore` command, but takes an additional `alias_name` parameter.
After the index has been uploaded to elasticsearch, any indexes which
were being addressed by the alias have the alias removed, the newly
uploaded index has the alias added, and finally any indexes which
were being addressed by the alias are closed.

This means that applications can perform searches by addressing the
index using the alias name, and from their point of view the uploaded
index replaces the old index atomically.