Add option to override settings when loading data
Add an optional parameter to the `restore` and `restore_alias` commands,
which takes a literal JSON string, and merges that with the
configuration passed when creating the index.

This allows some of the settings or mappings to be overridden when
loading the index, but uses the settings from the dump file by default.

I want to use this when taking a dump from a production system (which
has many shards and replicas) and loading it into a development
environment which has a single node.  For example:

```
    es_dump_restore restore_alias http://localhost:9200 test test-1276512 test_dump.zip \
    '{"settings":{"index":{"number_of_replicas":"0","number_of_shards":"1"}}'
```