Make the batch size configurable when restoring
Previously, the batch size when loading changes into elasticsearch was
fixed at 1000.  This commit changes the restore commands to take an
optional batch size parameter, as an integer, representing the number of
documents to send at once.

For GOV.UK, we've been experiencing timeouts for some of our more
complex documents, now that we have fairly complex mappings and analysis
chains for these.  This commit allows us to reduce the batch size, and
avoid the timeouts.